+++ crates/bcr-wallet-cli/src/command.rs
@@ -4,8 +4,8 @@ use anyhow::Result;
 use bcr_common::cdk_common::wallet::TransactionId;
 use bcr_wallet_api::{AppState, config::CreateWalletConfig};
 use bcr_wallet_core::types::{
-    PaymentResultCallback, TransactionFilters, TransactionSort, get_btc_alpha_tx_id,
-    get_btc_beta_tx_id, get_payment_type, get_transaction_status,
+    PaymentResultCallback, TransactionFilters, TransactionSort, get_btc_tx_id, get_payment_type,
+    get_transaction_status,
 };
 use chrono::{DateTime, Utc};
 use tokio::sync::oneshot;
@@ -88,19 +88,11 @@ pub async fn cmd_info(app_state: &AppState) -> Result<String> {
             for tx in transactions.iter() {
                 let status = get_transaction_status(&tx.metadata);
                 let ptype = get_payment_type(&tx.metadata);
-                let alpha_btc_tx_id = get_btc_alpha_tx_id(&tx.metadata);
-                let beta_btc_tx_id = get_btc_beta_tx_id(&tx.metadata);
-                let quote_or_btc_tx_id = match (beta_btc_tx_id, alpha_btc_tx_id, &tx.quote_id) {
-                    (Some(_), Some(_), Some(_)) => String::default(),
-                    (None, Some(_), Some(_)) => String::default(),
-                    (Some(_), None, Some(_)) => String::default(),
-                    (Some(beta_btc_tx_id), None, None) => beta_btc_tx_id.to_string(),
-                    (None, Some(alpha_btc_tx_id), None) => alpha_btc_tx_id.to_string(),
-                    (Some(beta_btc_tx_id), Some(alpha_btc_tx_id), None) => {
-                        format!("alpha: {}, beta: {}", alpha_btc_tx_id, beta_btc_tx_id)
-                    }
-                    (None, None, Some(quote_id)) => quote_id.to_string(),
-                    (None, None, None) => String::default(),
+                let btc_tx_id = get_btc_tx_id(&tx.metadata);
+                let quote_or_btc_tx_id = match (btc_tx_id, &tx.quote_id) {
+                    (Some(txid), _) => txid.to_string(),
+                    (None, Some(quote_id)) => quote_id.to_string(),
+                    (None, None) => String::default(),
                 };
                 res.push_str(&format!(
                     "\t\tId: {} \t Amount: {} {} \t Fees: {}  \t Status: {:?} \t {} \tType: {:<10} \t {:?} \t Memo: {} \t BTC TxID/Quote ID: {}",