+++ crates/bcr-wallet-ffi/src/frb_generated.rs
@@ -28,7 +28,7 @@
 
 use crate::api::*;
 use flutter_rust_bridge::for_generated::byteorder::{NativeEndian, ReadBytesExt, WriteBytesExt};
-use flutter_rust_bridge::for_generated::{transform_result_dco, Lifetimeable, Lockable};
+use flutter_rust_bridge::for_generated::{Lifetimeable, Lockable, transform_result_dco};
 use flutter_rust_bridge::{Handler, IntoIntoDart};
 
 // Section: boilerplate
@@ -2138,18 +2138,6 @@ impl SseDecode for Vec<crate::api::WalletDevModeDetailedBalanceEntry> {
     }
 }
 
-impl SseDecode for crate::api::MeltTx {
-    // Codec=Sse (Serialization based), see doc to use other codecs
-    fn sse_decode(deserializer: &mut flutter_rust_bridge::for_generated::SseDeserializer) -> Self {
-        let mut var_alphaTxId = <Option<String>>::sse_decode(deserializer);
-        let mut var_betaTxId = <Option<String>>::sse_decode(deserializer);
-        return crate::api::MeltTx {
-            alpha_tx_id: var_alphaTxId,
-            beta_tx_id: var_betaTxId,
-        };
-    }
-}
-
 impl SseDecode for crate::api::MigrateRabidResponse {
     // Codec=Sse (Serialization based), see doc to use other codecs
     fn sse_decode(deserializer: &mut flutter_rust_bridge::for_generated::SseDeserializer) -> Self {
@@ -2371,7 +2359,7 @@ impl SseDecode for crate::api::Transaction {
         let mut var_memo = <Option<String>>::sse_decode(deserializer);
         let mut var_ptype = <crate::api::PaymentType>::sse_decode(deserializer);
         let mut var_status = <crate::api::TransactionStatus>::sse_decode(deserializer);
-        let mut var_meltTx = <crate::api::MeltTx>::sse_decode(deserializer);
+        let mut var_txId = <Option<String>>::sse_decode(deserializer);
         let mut var_quoteId = <Option<String>>::sse_decode(deserializer);
         return crate::api::Transaction {
             id: var_id,
@@ -2383,7 +2371,7 @@ impl SseDecode for crate::api::Transaction {
             memo: var_memo,
             ptype: var_ptype,
             status: var_status,
-            melt_tx: var_meltTx,
+            tx_id: var_txId,
             quote_id: var_quoteId,
         };
     }
@@ -3324,22 +3312,6 @@ impl flutter_rust_bridge::IntoIntoDart<crate::api::IsValidTokenResponse>
     }
 }
 // Codec=Dco (DartCObject based), see doc to use other codecs
-impl flutter_rust_bridge::IntoDart for crate::api::MeltTx {
-    fn into_dart(self) -> flutter_rust_bridge::for_generated::DartAbi {
-        [
-            self.alpha_tx_id.into_into_dart().into_dart(),
-            self.beta_tx_id.into_into_dart().into_dart(),
-        ]
-        .into_dart()
-    }
-}
-impl flutter_rust_bridge::for_generated::IntoDartExceptPrimitive for crate::api::MeltTx {}
-impl flutter_rust_bridge::IntoIntoDart<crate::api::MeltTx> for crate::api::MeltTx {
-    fn into_into_dart(self) -> crate::api::MeltTx {
-        self
-    }
-}
-// Codec=Dco (DartCObject based), see doc to use other codecs
 impl flutter_rust_bridge::IntoDart for crate::api::MigrateRabidResponse {
     fn into_dart(self) -> flutter_rust_bridge::for_generated::DartAbi {
         [self.migrated_to_mint.into_into_dart().into_dart()].into_dart()
@@ -3556,7 +3528,7 @@ impl flutter_rust_bridge::IntoDart for crate::api::Transaction {
             self.memo.into_into_dart().into_dart(),
             self.ptype.into_into_dart().into_dart(),
             self.status.into_into_dart().into_dart(),
-            self.melt_tx.into_into_dart().into_dart(),
+            self.tx_id.into_into_dart().into_dart(),
             self.quote_id.into_into_dart().into_dart(),
         ]
         .into_dart()
@@ -4854,14 +4826,6 @@ impl SseEncode for Vec<crate::api::WalletDevModeDetailedBalanceEntry> {
     }
 }
 
-impl SseEncode for crate::api::MeltTx {
-    // Codec=Sse (Serialization based), see doc to use other codecs
-    fn sse_encode(self, serializer: &mut flutter_rust_bridge::for_generated::SseSerializer) {
-        <Option<String>>::sse_encode(self.alpha_tx_id, serializer);
-        <Option<String>>::sse_encode(self.beta_tx_id, serializer);
-    }
-}
-
 impl SseEncode for crate::api::MigrateRabidResponse {
     // Codec=Sse (Serialization based), see doc to use other codecs
     fn sse_encode(self, serializer: &mut flutter_rust_bridge::for_generated::SseSerializer) {
@@ -5047,7 +5011,7 @@ impl SseEncode for crate::api::Transaction {
         <Option<String>>::sse_encode(self.memo, serializer);
         <crate::api::PaymentType>::sse_encode(self.ptype, serializer);
         <crate::api::TransactionStatus>::sse_encode(self.status, serializer);
-        <crate::api::MeltTx>::sse_encode(self.melt_tx, serializer);
+        <Option<String>>::sse_encode(self.tx_id, serializer);
         <Option<String>>::sse_encode(self.quote_id, serializer);
     }
 }
@@ -5614,7 +5578,7 @@ mod io {
     use flutter_rust_bridge::for_generated::byteorder::{
         NativeEndian, ReadBytesExt, WriteBytesExt,
     };
-    use flutter_rust_bridge::for_generated::{transform_result_dco, Lifetimeable, Lockable};
+    use flutter_rust_bridge::for_generated::{Lifetimeable, Lockable, transform_result_dco};
     use flutter_rust_bridge::{Handler, IntoIntoDart};
 
     // Section: boilerplate
@@ -5653,7 +5617,7 @@ mod web {
     };
     use flutter_rust_bridge::for_generated::wasm_bindgen;
     use flutter_rust_bridge::for_generated::wasm_bindgen::prelude::*;
-    use flutter_rust_bridge::for_generated::{transform_result_dco, Lifetimeable, Lockable};
+    use flutter_rust_bridge::for_generated::{Lifetimeable, Lockable, transform_result_dco};
     use flutter_rust_bridge::{Handler, IntoIntoDart};
 
     // Section: boilerplate