+++ lib/src/rust/frb_generated.dart
@@ -2322,18 +2322,6 @@ class RustLibApiImpl extends RustLibApiImplPlatform implements RustLibApi {
   }
 
   @protected
-  MeltTx dco_decode_melt_tx(dynamic raw) {
-    // Codec=Dco (DartCObject based), see doc to use other codecs
-    final arr = raw as List<dynamic>;
-    if (arr.length != 2)
-      throw Exception('unexpected arr length: expect 2 but see ${arr.length}');
-    return MeltTx(
-      alphaTxId: dco_decode_opt_String(arr[0]),
-      betaTxId: dco_decode_opt_String(arr[1]),
-    );
-  }
-
-  @protected
   MigrateRabidResponse dco_decode_migrate_rabid_response(dynamic raw) {
     // Codec=Dco (DartCObject based), see doc to use other codecs
     final arr = raw as List<dynamic>;
@@ -2517,7 +2505,7 @@ class RustLibApiImpl extends RustLibApiImplPlatform implements RustLibApi {
       memo: dco_decode_opt_String(arr[6]),
       ptype: dco_decode_payment_type(arr[7]),
       status: dco_decode_transaction_status(arr[8]),
-      meltTx: dco_decode_melt_tx(arr[9]),
+      txId: dco_decode_opt_String(arr[9]),
       quoteId: dco_decode_opt_String(arr[10]),
     );
   }
@@ -3673,14 +3661,6 @@ class RustLibApiImpl extends RustLibApiImplPlatform implements RustLibApi {
   }
 
   @protected
-  MeltTx sse_decode_melt_tx(SseDeserializer deserializer) {
-    // Codec=Sse (Serialization based), see doc to use other codecs
-    var var_alphaTxId = sse_decode_opt_String(deserializer);
-    var var_betaTxId = sse_decode_opt_String(deserializer);
-    return MeltTx(alphaTxId: var_alphaTxId, betaTxId: var_betaTxId);
-  }
-
-  @protected
   MigrateRabidResponse sse_decode_migrate_rabid_response(
     SseDeserializer deserializer,
   ) {
@@ -3876,7 +3856,7 @@ class RustLibApiImpl extends RustLibApiImplPlatform implements RustLibApi {
     var var_memo = sse_decode_opt_String(deserializer);
     var var_ptype = sse_decode_payment_type(deserializer);
     var var_status = sse_decode_transaction_status(deserializer);
-    var var_meltTx = sse_decode_melt_tx(deserializer);
+    var var_txId = sse_decode_opt_String(deserializer);
     var var_quoteId = sse_decode_opt_String(deserializer);
     return Transaction(
       id: var_id,
@@ -3888,7 +3868,7 @@ class RustLibApiImpl extends RustLibApiImplPlatform implements RustLibApi {
       memo: var_memo,
       ptype: var_ptype,
       status: var_status,
-      meltTx: var_meltTx,
+      txId: var_txId,
       quoteId: var_quoteId,
     );
   }
@@ -5044,13 +5024,6 @@ class RustLibApiImpl extends RustLibApiImplPlatform implements RustLibApi {
   }
 
   @protected
-  void sse_encode_melt_tx(MeltTx self, SseSerializer serializer) {
-    // Codec=Sse (Serialization based), see doc to use other codecs
-    sse_encode_opt_String(self.alphaTxId, serializer);
-    sse_encode_opt_String(self.betaTxId, serializer);
-  }
-
-  @protected
   void sse_encode_migrate_rabid_response(
     MigrateRabidResponse self,
     SseSerializer serializer,
@@ -5240,7 +5213,7 @@ class RustLibApiImpl extends RustLibApiImplPlatform implements RustLibApi {
     sse_encode_opt_String(self.memo, serializer);
     sse_encode_payment_type(self.ptype, serializer);
     sse_encode_transaction_status(self.status, serializer);
-    sse_encode_melt_tx(self.meltTx, serializer);
+    sse_encode_opt_String(self.txId, serializer);
     sse_encode_opt_String(self.quoteId, serializer);
   }